Add backward compatibility wrapper
